getuserstats function moved data and domain module because it will be used in the other modules
