+++ domain/src/main/java/com/muratdayan/domain/repository/UsersDomainRepository.kt
@@ -1,12 +1,10 @@
-package com.muratdayan.game.domain.repository
+package com.muratdayan.domain.repository
 
 import com.muratdayan.common.AppError
 import com.muratdayan.common.Result
-import com.muratdayan.domain.model.FriendsDataModel
 import com.muratdayan.domain.model.UserStatsModel
 import kotlinx.coroutines.flow.Flow
 
-interface GameRepository {
-
-
+interface UsersDomainRepository {
+    fun getUserStats(): Flow<Result<UserStatsModel, AppError>>
 }
