+++ domain/src/main/java/com/muratdayan/domain/usecase/GetFriendsDomainUseCase.kt
@@ -1,10 +1,10 @@
 package com.muratdayan.domain.usecase
 
-import com.muratdayan.domain.repository.DomainRepository
+import com.muratdayan.domain.repository.FriendsDomainRepository
 import javax.inject.Inject
 
 class GetFriendsDomainUseCase @Inject constructor(
-    private val domainRepository: DomainRepository
+    private val friendsDomainRepository: FriendsDomainRepository
 ){
-    operator fun invoke() = domainRepository.getFriends()
+    operator fun invoke() = friendsDomainRepository.getFriends()
 }
