+++ domain/src/main/java/com/muratdayan/domain/usecase/GetUserStatsDomainUseCase.kt
@@ -1,0 +1,10 @@
+package com.muratdayan.domain.usecase
+
+import com.muratdayan.domain.repository.UsersDomainRepository
+import javax.inject.Inject
+
+class GetUserStatsDomainUseCase @Inject constructor(
+    private val usersDomainRepository: UsersDomainRepository
+){
+    operator fun invoke() = usersDomainRepository.getUserStats()
+}
