--- feature/game/src/main/java/com/muratdayan/game/domain/usecase/GetUserStatsUseCase.kt
@@ -1,10 +1,0 @@
-package com.muratdayan.game.domain.usecase
-
-import com.muratdayan.game.domain.repository.GameRepository
-import javax.inject.Inject
-
-class GetUserStatsUseCase @Inject constructor(
-    private val gameRepository: GameRepository
-){
-    operator fun invoke() = gameRepository.getUserStats()
-}
