+++ feature/game/src/main/java/com/muratdayan/game/presentation/main/MainScreenViewModel.kt
@@ -1,11 +1,10 @@
 package com.muratdayan.game.presentation.main
 
-import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.muratdayan.common.Result
 import com.muratdayan.domain.usecase.GetFriendsDomainUseCase
-import com.muratdayan.game.domain.usecase.GetUserStatsUseCase
+import com.muratdayan.domain.usecase.GetUserStatsDomainUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.flow.Flow
@@ -13,18 +12,16 @@ import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.catch
-import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.flow.onStart
 import kotlinx.coroutines.flow.receiveAsFlow
-import kotlinx.coroutines.flow.take
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @HiltViewModel
 class MainScreenViewModel @Inject constructor(
-    private val getUserStatsUseCase: GetUserStatsUseCase,
     private val getFriendsDomainUseCase: GetFriendsDomainUseCase,
+    private val getUserStatsDomainUseCase: GetUserStatsDomainUseCase
 ): ViewModel() {
 
     private val _uiState = MutableStateFlow(MainScreenContract.UiState())
@@ -109,7 +106,7 @@ class MainScreenViewModel @Inject constructor(
     private fun getUserStats() {
         viewModelScope.launch {
             updateUiState { copy(isLoading = true) }
-            getUserStatsUseCase.invoke()
+            getUserStatsDomainUseCase.invoke()
                 .onStart {
                     updateUiState { copy(isLoading = true) }
                 }
